api/requirements.txt
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
scikit-image
imageio
plotly
opencv-python
waitress
smart_open[s3]
boto3
amazon-dax-client
